Add filter for the args sent to wp_get_sites()
This will allow for a bit more flexibility in the use of this plugin. Multi-network users can set $args['network_id'] to null to get all sites -- currently the plugin is a bit difficult to use if you have a lot of networks, because you have to visit each network admin separately. Others may also want to ignore some blogs or set other args in wp_get_sites()